+++ packages/react/src/evm/index.ts
@@ -1,9 +1,4 @@
 export {
-  ConnectEmbed,
-  type ConnectEmbedProps,
-} from "../wallet/ConnectWallet/Modal/ConnectEmbed";
-
-export {
   useIsWalletModalOpen,
   useSetIsWalletModalOpen,
 } from "./providers/wallet-ui-states-provider";
@@ -15,10 +10,7 @@ export {
   ConnectWallet,
   type ConnectWalletProps,
 } from "../wallet/ConnectWallet/ConnectWallet";
-export {
-  ConnectModalInline,
-  type ConnectModalInlineProps,
-} from "../wallet/ConnectWallet/Modal/ConnectModalInline";
+export { ConnectModalInline } from "../wallet/ConnectWallet/Modal/ConnectModalInline";
 
 export { NetworkSelector } from "../wallet/ConnectWallet/NetworkSelector";
 export type { NetworkSelectorProps } from "../wallet/ConnectWallet/NetworkSelector";
@@ -82,5 +74,6 @@ export {
   setWalletAnalyticsEnabled,
   CoreWallet,
   DefiWallet,
+  RabbyWallet,
   Coin98Wallet,
 } from "@thirdweb-dev/wallets";