+++ packages/react/src/wallet/wallets/rabby/RabbyConnectUI.tsx
@@ -1,0 +1,112 @@
+import { ConnectUIProps, useConnect } from "@thirdweb-dev/react-core";
+import { ConnectingScreen } from "../../ConnectWallet/screens/ConnectingScreen";
+import { useCallback, useEffect, useRef, useState } from "react";
+import { GetStartedScreen } from "../../ConnectWallet/screens/GetStartedScreen";
+import type { RabbyWallet } from "@thirdweb-dev/wallets";
+import { wait } from "../../../utils/wait";
+import { useTWLocale } from "../../../evm/providers/locale-provider";
+import { RabbyScan } from "./RabbyScan";
+
+export const RabbyConnectUI = (props: ConnectUIProps<RabbyWallet>) => {
+  const [screen, setScreen] = useState<"connecting" | "get-started">(
+    "connecting",
+  );
+  const locale = useTWLocale().wallets.rabbyWallet;
+  const { walletConfig, connected } = props;
+  const connect = useConnect();
+  const [errorConnecting, setErrorConnecting] = useState(false);
+  const hideBackButton = props.supportedWallets.length === 1;
+
+  const connectToExtension = useCallback(async () => {
+    try {
+      connectPrompted.current = true;
+      setErrorConnecting(false);
+      setScreen("connecting");
+      await wait(1000);
+      await connect(walletConfig);
+      connected();
+    } catch (e) {
+      setErrorConnecting(true);
+      console.error(e);
+    }
+  }, [connected, connect, walletConfig]);
+
+  const connectPrompted = useRef(false);
+  useEffect(() => {
+    if (connectPrompted.current) {
+      return;
+    }
+
+    const isInstalled = walletConfig.isInstalled
+      ? walletConfig.isInstalled()
+      : false;
+
+    // if loading
+    (async () => {
+      if (isInstalled) {
+        connectToExtension();
+      }
+
+      // if wallet is not injected
+      else {
+        // No Rabby mobile app yet
+        setScreen("get-started");
+      }
+    })();
+  }, [connectToExtension, walletConfig]);
+
+  if (screen === "connecting") {
+    return (
+      <ConnectingScreen
+        locale={{
+          getStartedLink: locale.getStartedLink,
+          instruction: locale.connectionScreen.instruction,
+          tryAgain: locale.connectionScreen.retry,
+          inProgress: locale.connectionScreen.inProgress,
+          failed: locale.connectionScreen.failed,
+        }}
+        errorConnecting={errorConnecting}
+        onGetStarted={() => {
+          setScreen("get-started");
+        }}
+        onRetry={connectToExtension}
+        hideBackButton={hideBackButton}
+        onBack={props.goBack}
+        walletName={walletConfig.meta.name}
+        walletIconURL={walletConfig.meta.iconURL}
+      />
+    );
+  }
+
+  if (screen === "get-started") {
+    return (
+      <GetStartedScreen
+        locale={{
+          scanToDownload: locale.getStartedScreen.instruction,
+        }}
+        walletIconURL={walletConfig.meta.iconURL}
+        walletName={walletConfig.meta.name}
+        chromeExtensionLink={walletConfig.meta.urls?.chrome}
+        googlePlayStoreLink={walletConfig.meta.urls?.android}
+        appleStoreLink={walletConfig.meta.urls?.ios}
+        onBack={props.goBack}
+      />
+    );
+  }
+
+  if (screen === "scanning") {
+    return (
+      <RabbyScan
+        onBack={props.goBack}
+        onConnected={props.connected}
+        onGetStarted={() => {
+          setScreen("get-started");
+        }}
+        hideBackButton={hideBackButton}
+        walletConfig={walletConfig}
+      />
+    );
+  }
+
+  return null;
+};