+++ packages/wallets/src/evm/connectors/injected/types.ts
@@ -53,6 +53,7 @@ type InjectedProviderFlags = {
   isOkxWallet?: true;
   isCoreWallet?: true;
   isDefiWallet?: true;
+  isRabbyWallet?: true;
   isCoin98Wallet?: true;
 };
 type InjectedProviders = InjectedProviderFlags & {