+++ packages/wallets/src/evm/index.ts
@@ -56,6 +56,9 @@ export { getInjectedCoreWalletProvider } from "./connectors/core-wallet/getInjec
 export { DefiWallet, type DefiWalletOptions } from "./wallets/defi-wallet";
 export { getInjectedDefiWalletProvider } from "./connectors/defi-wallet/getInjectedDefiWalletProvider";
 
+export { RabbyWallet, type RabbyOptions } from "./wallets/rabby";
+export { getInjectedRabbyProvider } from "./connectors/rabby/getInjectedRabbyProvider";
+
 export { Coin98Wallet, type Coin98Options } from "./wallets/coin98";
 export { getInjectedCoin98Provider } from "./connectors/coin98/getInjectedCoin98Provider";